requirements.txt
absl-py==1.4.0
ale-py==0.8.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
cvxpy==1.3.1
cycler==0.11.0
ecos==2.0.12
filelock==3.10.7
fonttools==4.39.4
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
gym==0.26.2
gym-notices==0.0.8
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
osqp==0.6.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
qdldl==0.1.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
scs==3.2.3
seaborn==0.12.2
simpy==4.0.1
six==1.16.0
stable-baselines==2.10.2
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.1.0
torch==2.0.0
torch-tb-profiler==0.4.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
